fix: New API since moving to cityquiz.io

+++ HowManyCities/Data/City.swift
@@ -158,8 +158,8 @@ struct City: Codable, Hashable {
     
     let coordinates = (0..<10).map { i -> MKMapPoint in
       let scaleFactor = (i.isOdd ? smoothness : 1) * scaleFactor
-      let xCoordinate = scaleFactor * cos(.pi/2.0 - i*angleAdjustment) + center.x
-      let yCoordinate = scaleFactor * sin(.pi/2.0 - i*angleAdjustment) + center.y
+      let xCoordinate = scaleFactor * cos(.pi/2.0 - CGFloat(i)*angleAdjustment) + center.x
+      let yCoordinate = scaleFactor * sin(.pi/2.0 - CGFloat(i)*angleAdjustment) + center.y
       
       return .init(x: xCoordinate, y: yCoordinate)
     }
